SEQUOIADBMAINSTREAM-2057:【BUG】【OLD】SequoiaDB的UINT类型参数设为负值时处理错误，并且日志打印不正确 设置负值时报错及修改补充文档
git-svn-id: http://192.168.20.11/sequoiadb/trunk@36210 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/pmd/pmdEDUEntryPoint.cpp
@@ -74,7 +74,7 @@ namespace engine
    INT32 pmdSignalTestEntryPoint( pmdEDUCB *cb, void *arg )
    {
       pmdEDUCB *mainCB = ( pmdEDUCB* )arg ;
-      INT32 interval = pmdGetOptionCB()->getSignalInterval() ;
+      UINT32 interval = pmdGetOptionCB()->getSignalInterval() ;
       UINT32 timeCounter = 0 ;
 
       while( !cb->isDisconnected() )
@@ -83,7 +83,7 @@ namespace engine
          ++timeCounter ;
          interval = pmdGetOptionCB()->getSignalInterval() ;
 
-         if ( interval > 0 && timeCounter > (UINT32)interval )
+         if ( interval > 0 && timeCounter > interval )
          {
             ossPThreadKill( mainCB->getThreadID(), OSS_TEST_SIGNAL ) ;
             timeCounter = 0 ;
